pyproject.toml
[project]
name = "sponsoredai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.0",
    "streamlit>=1.30.0",
    "langchain>=0.1.0",
    "langchain-groq>=0.1.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "python-dotenv>=1.0.0",
    "plotly>=6.6.0",
    "pandas>=2.3.3",
    "zernio-sdk>=1.0.0",
    "requests>=2.31.0",
]
